requirements.txt
# Core dependencies for AI Video Generation Tool

# PDF Processing
PyPDF2==3.0.1
pdfplumber==0.10.3

# Text-to-Speech
gTTS==2.5.0
edge-tts==6.1.9

# Video Processing
moviepy==1.0.3
opencv-python==4.8.1.78
Pillow==10.2.0

# AI and NLP
openai==1.3.0
anthropic==0.7.0

# Audio Processing
pydub==0.25.1

# Utilities
python-dotenv==1.0.0
requests==2.31.0
tqdm==4.66.1
colorama==0.4.6

# Configuration
pyyaml==6.0.1

 copilot/add-report-and-presentation-tools
# Document and Presentation Generation
python-pptx==0.6.23
python-docx==1.1.0
openpyxl==3.1.2
reportlab==4.0.7

# Web Framework
flask==3.1.2
flask-cors==6.0.1

# Presentation Generation
python-pptx==1.0.2
 main
